Rename file name to show-banner.ts
