+++ client/jetpack-cloud/sections/utils/show-banner.ts
